+++ tests/e2e/user/user-registration.test.js
@@ -84,6 +84,7 @@ describe("A - Sign up and log in", () => {
             .end((err, res) => {
                 expect(res).to.have.status(200);
                 expect(res.body.email).to.equal(credentials.email);
+                expect(res.body.registration.method).to.equal("manual");
                 done();
             });
     });